refactor(api): strip cache + singleflight from /v1/admin/query
Now that raw SQL is an admin-only escape hatch (not the primary read
path), the L1/L2 + singleflight + mutation-cache-bypass machinery in
QueryHandler is dead weight. Admin calls are infrequent and often
mutations or one-shot — the cache hit rate is effectively zero and there
is no concurrent-request herd for singleflight to collapse. Every
request goes straight to ClickHouse.

QueryHandler shrinks to { CHConn driver.Conn }; Handle is
parse → executeQuery → marshal → write. NewQueryHandler(conn) replaces
the old (conn, *cache.TieredCache, time.Duration) signature; main.go
wiring follows.

The X-Cache: HIT/MISS header is gone from /v1/admin/query — every
response is a fresh round-trip, no header needed to disambiguate.
The structured-query and pipes endpoints keep their X-Cache (and the
full caching machinery) — those are the high-QPS read paths.

The verb-classification machinery (isMutation, mutationVerbs, the
comment-aware CTE parser) stays: it's correctness, not performance.
clickhouse-go's driver.Query() errors on statements that return no
result set, so we still need to route TRUNCATE/INSERT/DELETE/etc.
through driver.Exec.

The shared queryCacheKey helper moves to a new internal/api/cache_key.go
since it's now consumed only by structured_query.go and pipes.go.
TestQueryCacheKey_Deterministic moves with it.

TestQueryHandler_MutationBypassesCache is gone — no cache to bypass.
The mutation-routing tests (TestExecuteQuery_MutationRoutesToExec,
TestExecuteQuery_SelectRoutesToQuery) stay; they pin the still-required
Exec-vs-Query classification.
