+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- **`POST /v1/admin/query` no longer caches, singleflights, or bypasses cache for mutations** (`internal/api/query.go`, `internal/api/query_test.go`, `internal/api/cache_key.go` (new), `internal/api/cache_key_test.go` (new), `cmd/wavehouse/main.go`, `docs/src/content/docs/api.md`, `docs/src/content/docs/architecture.md`). With the endpoint now an admin-only escape hatch (see the Security entry below), the L1/L2 + singleflight + mutation-cache-bypass machinery added in the original raw-SQL handler is dead weight: admin calls are infrequent, often mutations, often one-shot, so the cache hit rate is effectively zero and singleflight has no concurrent-request herd to collapse. `QueryHandler` shrinks to `{ CHConn driver.Conn }`; `Handle` is parse → executeQuery → marshal → write. `NewQueryHandler(conn)` is the new signature (was `NewQueryHandler(conn, *cache.TieredCache, time.Duration)`). The `X-Cache: HIT/MISS` response header is gone from this endpoint — every response is a fresh ClickHouse round-trip, no header needed to disambiguate. The verb-classification machinery (`isMutation`, `mutationVerbs`, the comment-aware CTE parser) STAYS; it's correctness, not performance — clickhouse-go's `driver.Query()` errors on statements that return no result set, so we still route mutations through `driver.Exec`. The shared `queryCacheKey` helper moved to a new `internal/api/cache_key.go` since it's now consumed only by the structured query (`/v1/tables/{table}/query`) and pipes (`/v1/pipes/{name}`) handlers, which keep their caching + singleflight intact — those are the high-QPS read paths where the machinery earns its keep. Two now-obsolete tests removed: `TestQueryCacheKey_Deterministic` (moved to a fresh `cache_key_test.go` since the helper still has callers) and `TestQueryHandler_MutationBypassesCache` (no cache to bypass). The mutation-routing tests (`TestExecuteQuery_MutationRoutesToExec`, `TestExecuteQuery_SelectRoutesToQuery`) stay — they pin the still-required Exec-vs-Query classification.
+
 ### Removed
 
 - **BREAKING: the ingest pipeline is now insert-only; the `action` field on `ingest.<table>` envelopes is gone from the wire format** (`internal/ingest/bento.go`, `internal/ingest/bento_test.go`, `tests/integration/ingest_test.go`, `tests/integration/dlq_test.go`, `docs/src/content/docs/architecture.md`, `docs/src/content/docs/api.md`, `AGENTS.md`): the worker accepts `{table_name, received_timestamp, data}` envelopes (the existing `EventMessage` shape) and bulk-INSERTs them. All non-insert mutations (`DELETE`/`UPDATE`/`TRUNCATE`/`DROP`/`ALTER`/`REPLACE`/…) must go through `POST /v1/admin/query` under the `admin` / `service` role (the raw-SQL endpoint moved under `/v1/admin/*` in a follow-up commit on this same Unreleased section — see the Security entry below; the `RawSQL: true` policy field is gone too). The rationale: our policy engine authorizes mutations by evaluating row/column rules against the *payload* of an operation, which works for inserts but not for predicate-driven mutations (we can't prove a `WHERE` clause is satisfiable only for rows the caller is allowed to touch, and `WHERE 1=1` would otherwise nuke a table). Rather than ship a partial enforcement story, the pipelined surface narrows to the shape policy can authorize today; `/v1/admin/query` becomes the single sanctioned admin-equivalent surface for everything else. The embedded NATS server runs with `DontListen: true` (`internal/mq/embedded.go`), so the only Publishers reachable on `ingest.>` are in-process Go code (today, only the HTTP `/v1/ingest/{table}` handler) — no perimeter-level action-rejection check is needed at the consumer. The `delete-envelope` DLQ shape and the `Wave-DLQ-Type` NATS header it required are gone with this — `dlq.<table>` now carries only insert-failure payloads (the inner data object), so DLQ consumers no longer need to discriminate. The `jsInput.chConn` and `jsInput.js` fields are dropped along with the `chConn driver.Conn` parameter on `StartIngestWorker`. Re-introducing pipelined / structured mutations is deferred until the policy engine can authorize predicates — track separately if/when scheduled.
